Fix trest memory allocation checks
This patch fixes memory allocation checks when allocating headers for
the request. There were no checks on the memory allocations and the
syntax for allocations was reversed for calloc based allocations.

For memory allocations done via realloc the newly allocated memory
wasn't zeroed out but the trest request execution expects the headers to
be null terminated.

Signed-off-by: Pranay Kr. Srivastava <pranay.srivastava@pantacor.com>