fix a bug where we see in strace a busyloop of selects and reads
We observe a busyloop of select on writefds followed by a read syscall
which does not succeed. Those seem to happen as part of an implicit handshake
flow that happens if a timeout occurs during initial handshake.

We fix that code that tried to ensure that such timeout would be handled
properly